Add: 支持面板下发的精确限速 speed_limit_bps（字节/秒）
- UserInfo 新增 speed_limit_bps，面板按 KB/s、MB/s 设置的限速不再被取整到 Mbps
- 旧面板不下发该字段时仍按 speed_limit(Mbps) 限速，保持兼容
- limiter 内部限速统一改用字节/秒；限速变化检测包含 speed_limit_bps

+++ limiter/dynamic.go
@@ -1,7 +1,7 @@
 package limiter
 
-// determineSpeedLimit returns the minimum non-zero rate
-func determineSpeedLimit(limit1, limit2 int) (limit int) {
+// determineSpeedLimit returns the minimum non-zero rate (both in bytes/s)
+func determineSpeedLimit(limit1, limit2 int64) (limit int64) {
 	if limit1 == 0 || limit2 == 0 {
 		if limit1 > limit2 {
 			return limit1
